+++ packages/tsurlfilter/test/rules/declarative-converter/declarative-rule-converter.test.ts
@@ -685,11 +685,25 @@ describe('DeclarativeRuleConverter', () => {
                 condition: {
                     isUrlFilterCaseSensitive: false,
                     urlFilter: '||example.com',
+                    resourceTypes: [
+                        'main_frame',
+                        'sub_frame',
+                        'stylesheet',
+                        'script',
+                        'image',
+                        'font',
+                        'object',
+                        'xmlhttprequest',
+                        'ping',
+                        'media',
+                        'websocket',
+                        'other',
+                    ],
                 },
             });
         });
 
-        it('converts empty $removeparam rule', async () => {
+        it('converts $removeparam rule without parameters', async () => {
             const filterId = 0;
             const filter = await createFilter(
                 filterId,
@@ -714,6 +728,20 @@ describe('DeclarativeRuleConverter', () => {
                 condition: {
                     isUrlFilterCaseSensitive: false,
                     urlFilter: '||example.com',
+                    resourceTypes: [
+                        'main_frame',
+                        'sub_frame',
+                        'stylesheet',
+                        'script',
+                        'image',
+                        'font',
+                        'object',
+                        'xmlhttprequest',
+                        'ping',
+                        'media',
+                        'websocket',
+                        'other',
+                    ],
                 },
             });
         });